requirements.txt
anthropic>=0.25.0
instaloader>=4.10
Pillow>=10.0
requests>=2.31
